Run coverage report only on Debug build
The coverage report only works on GCC, because I cannot figure out how to coax the clang toolset to omit the old format that codecov.io expects.  So make this one a Debug build to get an accurate coverage report.  Build in Release mode with clang in a separate job.